requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==23.2.0
Automat==22.10.0
Babel==2.10.3
bcrypt==3.2.2
bitarray==3.8.1
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
ckzg==2.1.7
click==8.1.6
cloud-init==25.2
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
constantly==23.10.4
cryptography==41.0.7
cytoolz==1.1.0
dbus-python==1.3.2
distro==1.9.0
distro-info==1.7+build1
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
frozenlist==1.8.0
hexbytes==1.3.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.7.1
netifaces==0.11.0
oauthlib==3.2.2
parsimonious==0.10.0
propcache==0.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.23.0
pycurl==7.45.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu5.2
pytz==2024.1
pyunormalize==17.0.0
PyYAML==6.0.1
regex==2026.4.4
requests==2.31.0
rich==13.7.1
rlp==4.1.0
service-identity==24.1.0
setuptools==68.1.2
six==1.16.0
systemd-python==235
toolz==1.1.0
Twisted==24.3.0
types-requests==2.33.0.20260402
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
web3==7.15.0
websockets==15.0.1
wheel==0.42.0
yarl==1.23.0
zope.interface==6.1
